pyproject.toml
[project]
name = "newsagent"
version = "0.1.0"
description = ""
authors = [
    {name = "Jomalang",email = "rhkddlr98@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "crewai[tools] (>=0.100.1,<0.101.0)",
    "langchain (>=0.3.18,<0.4.0)",
    "langchain-openai (>=0.1.7,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"

